move classes related to openai to a seperate file and fix
I had to do some horrendous way to fix it up but now it works! yipiee

the fix was to
